Moved the Diverges struct to its own file
